Move the PullDispatcher to the CO.CDP.MQ project
