+++ Libraries/CO.CDP.AwsServices/Sqs/SqsDispatcher.cs
@@ -1,5 +1,6 @@
 using Amazon.SQS;
 using Amazon.SQS.Model;
+using CO.CDP.MQ.Pull;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
