+++ Services/CO.CDP.Organisation.WebApi/Events/EventDeserializer.cs
@@ -1,5 +1,5 @@
 using System.Text.Json;
-using CO.CDP.AwsServices.Sqs;
+using CO.CDP.MQ.Pull;
 using static CO.CDP.Organisation.WebApi.Events.EventDeserializer.EventDeserializerException;
 
 namespace CO.CDP.Organisation.WebApi.Events;
